Extract exported function into a new file.
